Fix: Board diameter expecting a long instead of a double

+++ src/main/java/ar/edu/itba/ss/Board.java
@@ -9,8 +9,8 @@ public class Board {
 
     private final Set<Particle> particles;
 
-    public Board(long L, double obstacleRadius, Set<Particle> particles) {
-        this.radius = (double)L/2;
+    public Board(double L, double obstacleRadius, Set<Particle> particles) {
+        this.radius = L/2;
         this.obstacleRadius = obstacleRadius;
         this.particles = particles;
     }
